fix Organization entity & adjust controller

+++ src/main/java/com/example/demo/Organization/OrganizationRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.Organization;
 import com.example.demo.Organization.Organization;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Long> {
+public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
 
 }
